+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * Binary format to store books in a file.
 * `Book::toc()` returns a tree of table of contents.
+* Both `Page` and `Book` can include any number of metadata entries.